+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2017-08-30
+
+### Added
+
+- Added the option to create the start and the end object for the spawn of the enemies
+
+### Changed
+
+- Organized free and path cells in two different gamobjects at the instantiation
+
+### Fixed
+
+- Code is now more readable
+- Each path is now connect with the others
+
 ## [1.3.0] - 2017-08-26
 
 ### Added